+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Created service to fetch state from a list of chainIds
+- `LeaderSelector` service
 
 ### Changed
 - Created api price freshness settings